requirements.txt
Flask==1.0.2
Werkzeug==0.15.5
flask_cors==3.0.9 
gunicorn==19.9.0
nose==1.3.7
requests==2.21.0
argparse==1.4.0
flask-swagger-ui==3.20.9
itsdangerous==2.0.1
mysql-connector-python==8.0.31
validate_email==1.3
uuid==1.30